+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## [Unreleased]
 
+## [0.2.0] - 2024-02-03
+
 - Rename `Sidekiq::Rescue::DSL` to `Sidekiq::Rescue::Dsl`
 - Update the `delay` option to now accept a proc as an argument
 - Update dsl to accept a list of errors
@@ -13,5 +15,6 @@
 - Add documentation
 - Add CI
 
-[Unreleased]: https://github.com/moofkit/sidekiq-rescue/compare/v0.1.0...HEAD
+[Unreleased]: https://github.com/moofkit/sidekiq-rescue/compare/v0.2.0...HEAD
+[0.2.0]: https://github.com/moofkit/sidekiq-rescue/releases/tag/v0.2.0
 [0.1.0]: https://github.com/moofkit/sidekiq-rescue/releases/tag/v0.1.0